Move scripts/components into subdirectories
